Added player detection to NPCs
I added the player detection code along with a function to check if there are obstacles in the way of the npc's line of sight and if the player is too far away from the npc to see. In doing this I updated the collision layers for all nodes with colliders.