Added Guards, Fixed detection/rotation bug
Added guards that use the player detection script and found that the guards' and cameras' rotation was off by 90 degrees. I hard coded the detection script to check for the node name of CameraBody to solve this, I expect this bug to resurface, but it works for now.

The guards have a movement script using the tilemap to navigate, and will choose a new path if there is an obstacle. They also detect the player with the same script as the Camera.