Created Locked doors and Password Generator
I added locked doors that can be opened by entering the correct code. The player enters codes through a UI element that has a number pad. I also added a custom node to generate passwords, but it is not yet connected to the doors. Currently the passwords are hard coded.