Connected specific doors to specific computers and other small things
I conected locked doors to computers and indicated which doors were connected to which computers by shading both the locked doors and computers the same color. I connected the locked doors and computers through signals that are sent from each computer to their respective locked door. I also changed the vision settings so that when you have nightvision you cannot see the color the locked doors and computers or the flashlights of the guards and cameras, but when you are in dark mode you can see all of these things. This was acheived by using groups and a dark_mode and nightvision_mode function. The braodcast is sent from the canvas layer that controlls the contrast.