Loot, gameover, victory and disguises implementation
In this I added a lot of things because I had committed for a while. First there is now loot in the game. moving next to the loot will pick it up and be displayed on the GUI.
You can disguise yourself to become "invincible" meaning the guards cant see you, but with some limitations. This has also been hooked into the GUI.
there is an exit now that will end the game if you go into the exit after having picked up the loot.
Instead of quitting the application you are new taken to either a gameover screen or victory screen depending on if you won or lost.

At some point the locked door system broke, so that only a null value is the correct answer and generated codes do not get connected to the doors, this requires some bugfixing.