Rough beta build
This is a full version of the game. It has a couple of bugs that can be fixed but by and large it is feature complete.
This commit added a main menu screen that will open when the game is run. From there you can decide what level to play. When you win or lose the game you a new option shows up in the game over/ you win scree that will take you back to the main menu. This project is going to be shelved for a little bit while I move to the 3D section of the Udemy course.